bug #968 Add accept header for the checks API (Agares)
This PR was squashed before being merged into the 3.0.x-dev branch.

Discussion
----------

Fixes #967.

This PR adds the Accept header back for the Checks API. This is required to make this work on GitHub Enterprise 2.22 (the latest stable version), which still considers this API a preview.

Commits
-------

d31106835197fd13904cb398a067cc476a08c721 Add accept header for the checks API
2ab07f34cb4de2a5cca72a74144a1af62c23f809 Change to the old way of setting headers